Enhancements & added a behaviour to retry the prompt if some went wrong

+++ common/domain/src/main/java/com/devalr/domain/repositories/impl/GeminiRepositoryImpl.kt
@@ -17,13 +17,16 @@ class GeminiRepositoryImpl(
             databasePrompt.promptResult
         } else {
             val promptResult = datasource.generateDaySummary(dataForPrompt = dataForPrompt)
-            database.removeAllDailyPrompts()
-            database.insertDailyPromptResult(
-                PromptResultEntity(
-                    date = getDate(),
-                    promptResult = promptResult
+            if (promptResult.isBlank().not()) {
+                database.removeAllDailyPrompts()
+                database.insertDailyPromptResult(
+                    PromptResultEntity(
+                        date = getDate(),
+                        promptResult = promptResult
+                    )
                 )
-            )
+            }
+
             promptResult
         }
     }
